requirements.txt
numpy==1.21.4
gym~=0.19.0
vizdoom==1.1.11
aenum==3.1.5
scikit-image~=0.15.0
tensorflow-gpu~=2.7.0
scipy~=1.7.3
matplotlib~=3.5.0
torch~=1.10.0